refactor(#148): logging 시스템 폴더 리팩토링
