+++ common/src/main/java/dukku/common/global/logging/kafaka/KafkaTracingBatchInterceptor.java
@@ -1,4 +1,4 @@
-package dukku.common.global.logging;
+package dukku.common.global.logging.kafaka;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.consumer.Consumer;
