+++ common/src/main/java/dukku/common/global/logging/logback/MdcTraceFilter.java
@@ -1,4 +1,4 @@
-package dukku.common.global.logging;
+package dukku.common.global.logging.logback;
 
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import ch.qos.logback.core.filter.Filter;
