+++ common/src/main/java/dukku/common/global/logging/logback/ServiceNameKeyingStrategy.java
@@ -1,4 +1,4 @@
-package dukku.common.global.logging;
+package dukku.common.global.logging.logback;
 
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import com.github.danielwegener.logback.kafka.keying.KeyingStrategy;
